fix(core): widen CoreError::invalid_handle parameter from u32 to u64
All handle APIs use u64, but invalid_handle() accepted u32, causing
every call site to truncate with `as u32`. This would produce wrong
error messages for handles exceeding u32::MAX.

Changed the parameter to u64 and removed all 8 truncation casts across
stream.rs (5), session.rs (2), and server.rs (1).

Closes #35

+++ crates/iroh-http-core/src/lib.rs
@@ -101,7 +101,7 @@ impl CoreError {
             message: detail.to_string(),
         }
     }
-    pub fn invalid_handle(handle: u32) -> Self {
+    pub fn invalid_handle(handle: u64) -> Self {
         CoreError {
             code: ErrorCode::InvalidInput,
             message: format!("unknown handle: {handle}"),
